Hide Linzer::Request module, rename it and move it to tests
This module is likely to confuse users and doesn't
add value really. It is only useful in the context
of unit tests, so let's keep it accessible only for
testing for now.
