+++ lib/linzer.rb
@@ -9,7 +9,6 @@ require "net/http"
 
 require_relative "linzer/version"
 require_relative "linzer/common"
-require_relative "linzer/request"
 require_relative "linzer/message"
 require_relative "linzer/message/adapter"
 require_relative "linzer/message/wrapper"
@@ -30,10 +29,6 @@ module Linzer
 
   class << self
     include Key::Helper
-
-    def new_request(verb, uri = "/", params = {}, headers = {})
-      Linzer::Request.build(verb, uri, params, headers)
-    end
 
     def verify(pubkey, message, signature, no_older_than: nil)
       Linzer::Verifier.verify(pubkey, message, signature, no_older_than: no_older_than)
