+++ spec/hmac_sha256_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe Linzer::Signer do
     let(:request) do
       request_data = Linzer::RFC9421::Examples.test_request_data
       path = request_data[:http]["path"]
-      Linzer.new_request(:post, path, {}, request_data[:headers])
+      Linzer::Test::Request.new_request(:post, path, {}, request_data[:headers])
     end
 
     let(:test_shared_secret_key_material) do
@@ -56,7 +56,7 @@ RSpec.describe Linzer::Verifier do
     let(:request) do
       request_data = Linzer::RFC9421::Examples.test_request_data
       path = request_data[:http]["path"]
-      Linzer.new_request(:post, path, {}, request_data[:headers])
+      Linzer::Test::Request.new_request(:post, path, {}, request_data[:headers])
     end
 
     let(:test_shared_secret_key_material) do
